split up jepa into base, ijepa, and iwm
