+++ cyto_dl/models/jepa/__init__.py
@@ -1,0 +1,1 @@
+from .jepa_base import JEPABase
